refactor: 테스트 전용 sql 설정
