+++ src/test/java/cart/controller/integration/ProductApiControllerTest.java
@@ -21,6 +21,7 @@ import org.springframework.http.MediaType;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
+import org.springframework.test.context.jdbc.Sql;
 
 import javax.sql.DataSource;
 
@@ -31,6 +32,7 @@ import static org.hamcrest.Matchers.equalTo;
 @DisplayNameGeneration(ReplaceUnderscores.class)
 @SuppressWarnings("NonAsciiCharacters")
 @SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
+@Sql("classpath:data-test.sql")
 class ProductApiControllerTest {
 
     private static final SqlParameterSource PARAMS = new MapSqlParameterSource()
