+++ src/test/java/cart/domain/EmailTest.java
@@ -1,6 +1,5 @@
-package cart.domain.member;
+package cart.domain;
 
-import cart.domain.Email;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
